+++ test/typechecking/unsupported/value_first_class_module.mli
@@ -1,0 +1,5 @@
+module type S = sig
+  val s : string
+end
+
+val f : (module S) -> string