+++ test/typechecking/unsupported/value_first_class_module_with_spec.mli
@@ -1,0 +1,12 @@
+module type S = sig
+  val s : string
+end
+
+val f : (module S) -> string
+(*@ s = f m *)
+(* {gospel_expected|
+[125] gospel: internal error, uncaught exception:
+              Gospel_checker.Uattr2spec.Unsupported
+              
+      
+|gospel_expected} *)